requirements.txt
pydantic>=2.7
pyyaml>=6.0
fastapi>=0.111
uvicorn[standard]>=0.30
qdrant-client>=1.9
openai>=1.40
sentence-transformers>=3.0
tenacity>=8.2
python-dotenv>=1.0
tiktoken>=0.7
httpx>=0.27
pytest>=8.2
pytest-cov>=5.0
reportlab>=4.2
python-docx>=1.1
docling>=2.0.0

